금액 관련 필드 Long 타입으로 교체

+++ app/src/main/java/com/toyprojects/card_pilot/model/Benefit.kt
@@ -2,7 +2,7 @@ package com.toyprojects.card_pilot.model
 
 data class Benefit(
     val name: String,
-    val used: Double,
-    val total: Double,
+    val used: Long,
+    val total: Long,
     val explanation: String? = null
 )
